service-os-agent/requirements.txt
langgraph>=0.3.0
langchain-anthropic>=0.3.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
supabase>=2.12.0
python-Levenshtein>=0.26.0
